fix: workdir

+++ apps/api/src/lib/common.ts
@@ -974,13 +974,14 @@ export const createDirectories = async ({
 }): Promise<{ workdir: string; repodir: string }> => {
 	const repodir = `/tmp/build-sources/${repository}/`;
 	const workdir = `/tmp/build-sources/${repository}/${buildId}`;
+	let workdirFound = false;
 	try {
-		await fs.stat(workdir);
+		workdirFound = !!(await fs.stat(workdir));
+	} catch (error) { }
+	if (workdirFound) {
 		await asyncExecShell(`rm -fr ${workdir}`);
-		await asyncExecShell(`mkdir -p ${workdir}`);
-	} catch(error) {
-		await asyncExecShell(`mkdir -p ${workdir}`);
 	}
+	await asyncExecShell(`mkdir -p ${workdir}`);
 	return {
 		workdir,
 		repodir
